// src/types/Role.ts
export enum Role {
  member_create = 'member_create',
  member_create_province = 'member_create_province',
  member_create_imi_paid = 'member_create_imi_paid',
  member_list_read = 'member_list_read',
  member_list_read_province = 'member_list_read_province',
  member_list_export = 'member_list_export',
  member_list_export_province = 'member_list_export_province',
  member_list_card_request = 'member_list_card_request',
  member_list_card_request_province = 'member_list_card_request_province',
  member_details_read = 'member_details_read',
  member_details_card_request = 'member_details_card_request',
  member_details_read_province = 'member_details_read_province',
  member_details_card_request_province = 'member_details_card_request_province',
  member_details_update = 'member_details_update',
  member_details_update_basic_profile = 'member_details_update_basic_profile',
  member_details_update_kta_number = 'member_details_update_kta_number',
  member_details_update_expiration_date = 'member_details_update_expiration_date',
  member_claim_list_read = 'member_claim_list_read',
  member_claim_list_update = 'member_claim_list_update',
  member_claim_list_read_province = 'member_claim_list_read_province',
  member_claim_list_update_province = 'member_claim_list_update_province',
  member_card_printing_request = 'member_card_printing_request',
  member_card_printing_read = 'member_card_printing_read',
  member_card_printing_process = 'member_card_printing_process',
  member_card_printing_download = 'member_card_printing_download',
  member_card_printing_read_province = 'member_card_printing_read_province',
  member_card_printing_process_province = 'member_card_printing_process_province',
  member_card_printing_download_province = 'member_card_printing_download_province',
  club_create = 'club_create',
  club_create_province = 'club_create_province',
  club_create_imi_paid = 'club_create_imi_paid',
  club_list_read = 'club_list_read',
  club_list_read_province = 'club_list_read_province',
  club_list_export = 'club_list_export',
  club_details_read = 'club_details_read',
  club_details_update = 'club_details_update',
  club_details_update_basic_profile = 'club_details_update_basic_profile',
  club_details_update_expiration_date = 'club_details_update_expiration_date',
  club_claim_list_read = 'club_claim_list_read',
  club_claim_list_update = 'club_claim_list_update',
  club_claim_list_read_province = 'club_claim_list_read_province',
  club_claim_list_update_province = 'club_claim_list_update_province',
  kis_configure_read = 'kis_configure_read',
  kis_configure_update = 'kis_configure_update',
  kis_pricing_read = 'kis_pricing_read',
  kis_pricing_range_update = 'kis_pricing_range_update',
  kis_pricing_read_province = 'kis_pricing_read_province',
  kis_pricing_update_province = 'kis_pricing_update_province',
  dashboard_read = 'dashboard_read',
  dashboard_read_province = 'dashboard_read_province',
  member_report_read = 'member_report_read',
  member_report_read_province = 'member_report_read_province',
  club_report_read = 'club_report_read',
  club_report_read_province = 'club_report_read_province',
}